refactor(config): externalize Kafka consumer props & lock down constants
- Move bootstrapServers and groupId into @ConfigurationProperties beans
  (LoginConsumerProperties, BookingConsumerProperties) annotated with @Component
- Remove manual registration of consumer property classes from main config
- Make NotificationLoginMessageKeys constructor private to prevent instantiation
