+++ notification-service/src/main/java/intexsoft/practice/notification_service/config/client/UserClientProperties.java
@@ -1,10 +1,12 @@
-package intexsoft.practice.notification_service.config;
+package intexsoft.practice.notification_service.config.client;
 
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
 
 @ConfigurationProperties(prefix = "user.client")
+@Component
 @Getter
 @Setter
 public class UserClientProperties {
