+++ notification-service/src/main/java/intexsoft/practice/notification_service/config/kafka/BaseConsumerProperties.java
@@ -1,0 +1,11 @@
+package intexsoft.practice.notification_service.config.kafka;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public abstract class BaseConsumerProperties {
+    private String bootstrapServers;
+    private String groupId;
+}
