+++ notification-service/src/main/java/intexsoft/practice/notification_service/config/kafka/BookingConsumerProperties.java
@@ -1,0 +1,13 @@
+package intexsoft.practice.notification_service.config.kafka;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@ConfigurationProperties(prefix = "consumer.booking")
+@Component
+@Getter
+@Setter
+public class BookingConsumerProperties extends BaseConsumerProperties {
+}
