+++ notification-service/src/main/java/intexsoft/practice/notification_service/localization/NotificationLoginMessageKeys.java
@@ -25,6 +25,6 @@ public final class NotificationLoginMessageKeys {
             UNKNOWN
     );
 
-    public NotificationLoginMessageKeys() {
+    private NotificationLoginMessageKeys() {
     }
 }
