+++ notification-service/src/main/resources/application.properties
@@ -9,6 +9,11 @@ spring.kafka.consumer.properties.spring.json.trusted.packages=intexsoft.practice
 spring.kafka.login.topic=auth.events.login
 spring.kafka.booking.topic=booking-created
 
+consumer.login.bootstrap-servers=localhost:9092
+consumer.login.group-id=login-service-group
+consumer.booking.bootstrap-servers=localhost:9092
+consumer.booking.group-id=booking-service-group
+
 spring.liquibase.change-log=classpath:db/changelog/db.changelog-master.xml
 
 spring.jpa.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
